Remove coloring from salt event summary output
Added options to the debug-salt script to allow using color or text for marking succeeded or failed events. Updated the call from supportconfig into the debug-salt script to use no color and to use text 'GOOD:' and 'FAIL:' to mark events that succeeded or failed, so that that output can in turn be displayed and utilized through CI which is just text and won't properly show the escape code colorization.

Additionally fixed a bug that was preventing the debug-salt from being run directly against an input XML fail using the '-i' option.

Added appropriate usage output to the debug-salt script when run directly with no options.